refactor(aw-export): move to `AntwebFormatter::ReferenceLink`
